requirements.txt
python-telegram-bot[job-queue]>=22.0,<23.0
markovify>=0.9.4,<1.0.0
google-genai>=1.30.0
groq>=0.9.0
python-dotenv>=1.0.1,<2.0.0
aiosqlite>=0.20.0,<1.0.0
